fix(ui-dashboard): cover OG + sUSDS queries in the GraphQL contract test
claude[bot] P1 / Codex P2: POOL_OG_DAILY_SNAPSHOTS, HOMEPAGE_OG_DAILY_SNAPSHOTS,
and SUSDS_YIELD_SUMMARY_QUERY hit real indexer fields (PoolDailySnapshot,
SusdsYieldSummary) but were module-private in lib/*-og.ts / reserve-yield-susds.ts,
so a field rename on the indexer would pass CI and fail only at runtime on the
OG image / reserve-yield routes.

Relocate the three query strings into the already-covered queries/ modules
(pools.ts, reserve-yield.ts) and import them back into the OG/sUSDS modules.
This is the repo convention (query strings live in queries/) and keeps the
exports knip-clean (used by their consumers) — cleaner than exporting from the
*-og.ts files, which import next/cache and can't load in the node test env.
The contract test now discovers 109 queries (was 106).

Also adds a comment noting the deliberate query-only scope (claude[bot] P3).

Refs #845
