+++ ui-dashboard/src/lib/__tests__/queries.test.ts
@@ -40,6 +40,8 @@ const EXPECTED_EXPORT_NAMES = [
   "BROKER_DAILY_SNAPSHOTS_ALL",
   "BROKER_EXCHANGE_DAILY_SNAPSHOTS_24H",
   "POOL_DAILY_FEE_SNAPSHOTS_PAGE",
+  "POOL_OG_DAILY_SNAPSHOTS",
+  "HOMEPAGE_OG_DAILY_SNAPSHOTS",
   "TRADING_LIMITS",
   "ORACLE_SNAPSHOTS",
   "ORACLE_SNAPSHOTS_CHART",
@@ -69,6 +71,7 @@ const EXPECTED_EXPORT_NAMES = [
   "CDP_TROVE_SCHEMA_FIELDS",
   "CDP_TROVE_OP_SNAPSHOTS",
   "SUSDS_YIELD_DAILY_SNAPSHOTS",
+  "SUSDS_YIELD_SUMMARY_QUERY",
   "POOL_BREAKER_CONFIG",
   "POOL_LABELS_ALL",
   "VIRTUAL_POOL_LIFECYCLE",
