+++ ui-dashboard/src/lib/queries/reserve-yield.ts
@@ -27,3 +27,24 @@ export const SUSDS_YIELD_DAILY_SNAPSHOTS = `
     }
   }
 `;
+
+// Consumed by lib/reserve-yield-susds.ts. Defined here so the GraphQL contract
+// test covers it (the contract test imports this queries module directly).
+export const SUSDS_YIELD_SUMMARY_QUERY = /* GraphQL */ `
+  query SusdsYieldSummary($id: String!) {
+    SusdsYieldSummary(where: { id: { _eq: $id } }, limit: 1) {
+      id
+      currentShares
+      costBasisUsdWei
+      realizedYieldUsdWei
+      transferredOutYieldUsdWei
+      redeemedYieldUsdWei
+      currentValueUsdWei
+      unrealizedYieldUsdWei
+      totalEarnedYieldUsdWei
+      sharePriceUsdWei
+      lastUpdatedBlock
+      lastUpdatedTimestamp
+    }
+  }
+`;
